requirements.txt
streamlit
google-generativeai
google-genai
PyMuPDF
reportlab
